third_party/blink/web_tests/platform/win/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/cast.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] cast float16 4D tensor to int64
  assert_less_than_equal: assert_array_approx_equals_ulp: test cast int64 actual 137438953475 should be close enough to expected 3 by ULP distance: expected a number less than or equal to 0n but got 137438953472n
[FAIL] cast int64 4D tensor to float16
  assert_less_than_equal: assert_array_approx_equals_ulp: test cast float16 actual 0 should be close enough to expected 1 by ULP distance: expected a number less than or equal to 1 but got 15360
Harness: the test ran to completion.
